fixed getting gpu use and cast

+++ radeon-profile/ioctlHandler.cpp
@@ -114,8 +114,8 @@ bool ioctlHandler::getGpuUsage(float *data, int time, int frequency) const {
 }
 
 
-bool ioctlHandler::getVramUsagePercentage(float *data) const {
-    float usage=0, total=0;
+bool ioctlHandler::getVramUsagePercentage(long *data) const {
+    long usage=0, total=0;
     bool success = getVramUsage(&usage) && getVramSize(&total);
     if(Q_LIKELY(success))
         *data = (100.0f * usage) / total;
